% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{as.data.frame.index}
\alias{as.data.frame.index}
\title{data frame conversion}
\usage{
\method{as.data.frame}{index}(x, row.names = NULL, optional = FALSE,
  ...)
}
\arguments{
\item{x}{an index object}

\item{row.names}{optional character vector of row names}

\item{optional}{logical. see ?as.data.frame}

\item{...}{additional parameters to be passed to functions}
}
\description{
data frame conversion
}
