% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{daysInMonth}
\alias{daysInMonth}
\title{daysInMonth}
\usage{
daysInMonth(x)
}
\arguments{
\item{x}{A date}
}
\value{
The number of days in the month in which x falls
}
\description{
calculate the number of days in the month that the given date falls
}
\keyword{internal}
