% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{dim.index}
\alias{dim.index}
\title{index dimension}
\usage{
\method{dim}{index}(x)
}
\arguments{
\item{x}{an index object}
}
\description{
Overrides the dim function so that it acts on the index element
of the index object
}
