% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{dimnames<-.index}
\alias{dimnames<-.index}
\title{index dimnames assignment}
\usage{
\method{dimnames}{index}(x) <- value
}
\arguments{
\item{x}{an index object}

\item{value}{value to assign}
}
\description{
Overrides the dimnames function to assign names to the index element
of the index class
}
