% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{dimnames.index}
\alias{dimnames.index}
\title{index dimnames}
\usage{
\method{dimnames}{index}(x)
}
\arguments{
\item{x}{an index object}
}
\description{
Overrides the dimnames function so that it acts on the index element
of the index object
}
