% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DominicksWeeks}
\alias{DominicksWeeks}
\title{Date information for the Dominicks data}
\format{
A data frame with 400 rows and 4 columns:
\describe{
  \item{week}{the number of the week}
  \item{start}{date the week started}
  \item{end}{date the week ended}
  \item{specialEvents}{special events, such as Halloween, that occurred during the week}
}
}
\source{
Dominicks Data Manual, Chicago Booth Kilts Center for Marketing, 2018, pages 21-28.
}
\usage{
DominicksWeeks
}
\description{
Table from the \href{https://www.chicagobooth.edu/-/media/enterprise/centers/kilts/datasets/dominicks-dataset/dominicks-manual-and-codebook_kiltscenter.aspx}{Dominicks Data Manual},
that gives the start and end date of each of the weeks in the movement
files.
}
\keyword{datasets}
