% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluateMatched.R
\name{evaluateMatched}
\alias{evaluateMatched}
\title{Evaluate product overlap between periods}
\usage{
evaluateMatched(x, pvar, qvar, pervar, prodID, output = "pop")
}
\arguments{
\item{x}{A dataframe containing price, quantity, a time period identifier
and a product identifier. It must have column names.}

\item{pvar}{A character string for the name of the price variable}

\item{qvar}{A character string for the name of the quantity variable}

\item{pervar}{A character string for the name of the time variable. This variable
must contain integers starting at period 1 and increasing in increments of 1 period.
There may be observations on multiple products for each time period.}

\item{prodID}{A character string for the name of the product identifier}

\item{output}{A character string specifying whether the matching should be
done assuming a chained index or a fixed base index. No index is actually computed,
but the matching needs to know which periods are being compared.}
}
\value{
A list of two matrices, one for expenditures and one for counts. Each
matrix has eight columns. The first four columns present the base period information
base_index (the index of the base period), base (base period expenditure or count),
base_matched (the expenditure or count of the base period after matching),
base_share (share of total expenditure in the base period that remains after
matching). Columns 5-8 are defined analagously for the current period. The matched
numbers for the base period should be interpreted as the count or expenditure
that remains after removal of products that exist in the base period, but not
in the current period. That is, products that existed in the base period but no
longer exist in the current period are removed by the matching. If new products
exist in the current period that were not available in the base period, this
does not affect the matched base period expenditure or count. The appearance
of new products is captured in the current period matched expenditure and counts.
Therefore, a base period share that is less than 1 indicates that products have
disappeared, while a current period share less than 1 indicates that new products
have appeared.
}
\description{
Evaluate the counts and expenditure for each period with and without
matching items across periods.
}
