% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bilateral.R
\name{fixed_t}
\alias{fixed_t}
\title{Fixed-basket indices - Laspeyres (q=q0), Paasche (q=q1), Lowe (q=qb)}
\usage{
fixed_t(p0, p1, q)
}
\description{
compute a bilateral fixed base index for a single period
}
\keyword{internal}

