% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilateral.R
\name{GEKS_w}
\alias{GEKS_w}
\title{GEKS_w}
\usage{
GEKS_w(x, pvar, qvar, pervar, indexMethod = "tornqvist", prodID,
  sample = "matched")
}
\description{
Function to compute a GEKS index over a window.
The window size is assumed to be the number of periods available in pervar.
This is not exposed to the user because GEKSIndex calls this
}
\keyword{internal}
