% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getFrequency}
\alias{getFrequency}
\title{getFrequency}
\usage{
getFrequency(x)
}
\arguments{
\item{x}{a vector of dates}
}
\value{
an integer indicating the number of
periods per year.
}
\description{
get the frequency of a vector of dates
}
\keyword{internal}
