% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gk.R
\name{GKIndex}
\alias{GKIndex}
\title{Compute the Geary-Khamis index}
\usage{
GKIndex(x, pvar, qvar, pervar, prodID, window, splice = "mean")
}
\arguments{
\item{x}{A dataframe containing price, quantity, a time period identifier
and a product identifier. It must have column names.}

\item{pvar}{A character string for the name of the price variable}

\item{qvar}{A character string for the name of the quantity variable}

\item{pervar}{A character string for the name of the time variable. This variable
must contain integers starting at period 1 and increasing in increments of 1 period.
There may be observations on multiple products for each time period.}

\item{prodID}{A character string for the name of the product identifier}

\item{window}{An integer specifying the length of the GEKS window.}

\item{splice}{the splicing method to use to extend the index. Valid methods are
window, movement, half or mean. The default is mean.}
}
\description{
Compute the Geary-Khamis index
}
