% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{index}
\alias{index}
\title{Create a new object of class 'index'}
\usage{
index(x, method, sample, output, chainMethod = "", sigma = NA)
}
\arguments{
\item{x}{an n by 1 matrix containing the price or quantity index}

\item{method}{the index number method used to compute the
index}

\item{sample}{whether matching was requested}

\item{output}{whether period-on-period or chained output was chosen}

\item{chainMethod}{the method of chaining chosen. The default
is "" if the chainMethod is not specified}

\item{sigma}{the value of sigma for a CES index, default is NA}
}
\description{
Create a new object of class 'index'
}
