% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndexNumR_index.R
\name{indexData}
\alias{indexData}
\title{IndexData}
\usage{
indexData(x)
}
\arguments{
\item{x}{an IndexNumR_Index}
}
\description{
Gets the index data as a numeric vector without the IndexNumR attributes
}
