% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticity.R
\name{indexDiff}
\alias{indexDiff}
\title{Computes the difference between a CES index and a comparison index}
\usage{
indexDiff(sigma, p0, p1, q0, q1, compIndex = "ces")
}
\description{
A function, with a root at zero, that computes the difference between
the CES index and a comparison index.
}
\keyword{internal}
