% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{isContinuous}
\alias{isContinuous}
\title{isContinuous}
\usage{
isContinuous(x)
}
\arguments{
\item{x}{Vector to check}
}
\value{
a list where the first element contains the
result of the check and the second element contains
the list of missing elements.
}
\description{
checks if a numeric vector has gaps.
}
\keyword{internal}
