% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IsraeliCPI}
\alias{IsraeliCPI}
\title{Data on fruits and vegetables from the Israeli CPI}
\format{
A list of two dataframes, one for fruits and one for vegetables.
Each of the dataframes has 6 variables and 504 observations (72 observations for each product)
\describe{
  \item{date}{the end-of-month date for the year and month of the data. In the tables
  in Diewert, Finkel and Artsev (2009) only the year and month are given.}
  \item{month}{the number of the month from 1-72}
  \item{prodID}{the name of the product}
  \item{price}{the price of the product for the given year and month}
  \item{quantity}{the quantity of the product sold. Calculated as expenditure/price}
  \item{expenditures}{the expenditure on the product for each year and month}
}
}
\source{
W. Erwin Diewert, Yoel Finkel and Yevgeny Artsev (2009), "Empirical Evidence on the
Treatment of Seasonal Products: The Israeli CPI Experience", chapter 4, pp. 53-78 in
W.E. Diewert, B.M. Balk, D. Fixler, K.J. Fox and A.O. Nakamura (eds.),
PRICE AND PRODUCTIVITY MEASUREMENT: Volume 2 -- Seasonality. Trafford Press.
Also available as a free e-publication at \url{http://www.indexmeasures.ca}.
}
\usage{
IsraeliCPI
}
\description{
This combines tables in Appendix B of Diewert, Finkel and Artsev (2009). Prices and expenditures
are taken directly from the tables. Other columns are constructed as detailed in the format section.
}
\keyword{datasets}
