% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bilateral.R
\name{lloydMoulton_tc}
\alias{lloydMoulton_tc}
\title{Lloyd-Moulton index, current period share}
\usage{
lloydMoulton_tc(p0, p1, q, sigma)
}
\description{
Lloyd-Moulton index, current period share
}
\keyword{internal}
