% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{Math.index}
\alias{Math.index}
\title{index Math}
\usage{
\method{Math}{index}(x, ...)
}
\arguments{
\item{x}{an index object}

\item{...}{other arguments}
}
\description{
Overrides the math functions so that they act on the index element
of the index object. For affected functions see ?groupGeneric.
}
