% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitValues.R
\name{monthIndex}
\alias{monthIndex}
\title{Generate an index of months}
\usage{
monthIndex(x)
}
\arguments{
\item{x}{A vector or column of dates}
}
\description{
A function to create a month index variable
}
\examples{
# given a vector of dates
df <- data.frame(date = as.Date(c("2017-01-01","2017-02-01","2017-03-01","2017-04-01"),
format = "\%Y-\%m-\%d"))
# calculate the time period variable
df$period <- monthIndex(df$date)
df
}
