% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{new_index}
\alias{new_index}
\title{default index object constructor}
\usage{
new_index(x, method, sample, output, chainMethod = "", sigma = NA)
}
\arguments{
\item{x}{an n by 1 matrix containing the price or quantity index}

\item{method}{the index number method used to compute the
index}

\item{sample}{whether matching was requested}

\item{output}{whether period-on-period or chained output was chosen}

\item{chainMethod}{the method of chaining chosen.  The default
is "" if the chainMethod is not specified}

\item{sigma}{the value of sigma for a CES index, default is NA}
}
\description{
Creates a new index object. Not exported; this is exposed by index().
}
\keyword{internal}
