% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{Ops.index}
\alias{Ops.index}
\title{index Ops}
\usage{
\method{Ops}{index}(e1, e2)
}
\arguments{
\item{e1}{an index object}

\item{e2}{an index object}
}
\description{
Overrides the Ops functions so that they act on the index element
of the index object. For affected functions see ?groupGeneric.
}
