% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{print.index}
\alias{print.index}
\title{index print}
\usage{
\method{print}{index}(x, ...)
}
\arguments{
\item{x}{an index object}

\item{...}{other arguments}
}
\description{
Overrides the print function so that it acts on the index element
of the index object
}
