% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{[.index}
\alias{[.index}
\title{index subsetting}
\usage{
\method{[}{index}(x, i, j, ...)
}
\arguments{
\item{x}{an index object}

\item{i}{row number}

\item{j}{column number}

\item{...}{other parameters}
}
\description{
Overrides the subset behaviour so that it acts on the index element
of the index object
}
