% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoStep.R
\name{subgroupPriceIndexes}
\alias{subgroupPriceIndexes}
\title{Calculate a price index on subgroups}
\usage{
subgroupPriceIndexes(subgroup, indexFunction, indexArgs)
}
\arguments{
\item{subgroup}{the name of the variable containing the subgroup ID. This
must be a factor variable, or a variable coercible to a factor.}

\item{indexFunction}{the function to use to calculate the index. Available
options are `priceIndex`, `GEKSIndex`, `GKIndex`, `WTPDIndex`.}

\item{indexArgs}{arguments for the price index function as a named list.
All arguments must be named.}
}
\description{
Calculate a price index on subgroups
}
