% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{[<-.index}
\alias{[<-.index}
\title{index subset assign}
\usage{
\method{[}{index}(x, i, j, ...) <- value
}
\arguments{
\item{x}{an index object}

\item{i}{row number}

\item{j}{column number}

\item{...}{other parameters}

\item{value}{value to be assigned}
}
\description{
Overrides the subset assign behaviour so that it acts on the index element
of the index object
}
