% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{Summary.index}
\alias{Summary.index}
\title{index Summary functions}
\usage{
\method{Summary}{index}(x, ..., na.rm = FALSE)
}
\arguments{
\item{...}{further arguments}

\item{na.rm}{logical: should missing values be removed? Default = FALSE}
}
\description{
Overrides the Summary group generic functions so that they act on
the index element of the index object. For affected functions see ?groupGeneric.
}
