% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{t.index}
\alias{t.index}
\title{index transpose}
\usage{
\method{t}{index}(x)
}
\arguments{
\item{x}{an index object}
}
\description{
Overrides the transpose function so that it acts on the index element
of the index object
}
