% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoStep.R
\name{twoStepIndex}
\alias{twoStepIndex}
\title{calculate a two-step index}
\usage{
twoStepIndex(
  x,
  pvar,
  qvar,
  pervar,
  prodID,
  subgroup,
  output = "pop",
  stepOneArgs = list(),
  stepTwoArgs = list()
)
}
\arguments{
\item{x}{A dataframe containing price, quantity, a time period identifier
and a product identifier. It must have column names.}

\item{pvar}{A character string for the name of the price variable}

\item{qvar}{A character string for the name of the quantity variable}

\item{pervar}{A character string for the name of the time variable. This variable
must contain integers starting at period 1 and increasing in increments of 1 period.
There may be observations on multiple products for each time period.}

\item{prodID}{A character string for the name of the product identifier}

\item{subgroup}{the name of the variable containing the subgroup ID. This
must be a factor variable, or a variable coercible to a factor.}

\item{output}{A character string specifying whether a chained (output="chained")
, fixed base (output="fixedbase") or period-on-period (output="pop")
price index numbers should be returned. Default is period-on-period.}

\item{stepOneArgs}{arguments to be passed to the price index function
in the first step as a named list. All arguments must be named.}

\item{stepTwoArgs}{arguments to be passed to the price index function
in the second step as a named list. All arguments must be named.}
}
\description{
calculate a two-step index
}
