% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/values.R
\name{values}
\alias{values}
\title{Compute values (price x quantity)}
\usage{
values(
  x,
  pvar,
  qvar,
  pervar,
  prodID,
  sample = "matched",
  matchPeriod = "previous",
  basePeriod = 1,
  valuePeriod = "current"
)
}
\arguments{
\item{x}{A dataframe containing price, quantity, a time period identifier
and a product identifier. It must have column names.}

\item{pvar}{A character string for the name of the price variable}

\item{qvar}{A character string for the name of the quantity variable}

\item{pervar}{A character string for the name of the time variable. This variable
must contain integers starting at period 1 and increasing in increments of 1 period.
There may be observations on multiple products for each time period.}

\item{prodID}{A character string for the name of the product identifier}

\item{sample}{A character string specifying whether a matched sample
should be used.}

\item{matchPeriod}{A character string specifying which period is used
to determine the set of products used for matching. Options are
"following", "previous", "base" or "current". expenditures are calculated
for the valuePeriod, after filtering out products that do not occur in
the "following", "previous", "base" or "current" periods respectively. The
terms "following" and "previous" should be interpreted as relative to the
current period.}

\item{basePeriod}{the base period to use when matchPeriod = "base".
Default is the first period.}

\item{valuePeriod}{A character string specifying which period's values
to return with matching applied if requested. "current" returns the
current period values, "previous" returns the previous period's values, and
"base" returns the base period values. Default is "current".}
}
\description{
Compute the total value (expenditure), for each time period in
the sample.
}
\examples{
values(CES_sigma_2, pvar = "prices", qvar = "quantities", pervar = "time",
prodID = "prodID", matchPeriod = "previous")

}
