% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitValues.R
\name{yearIndex}
\alias{yearIndex}
\title{Generate an index of years}
\usage{
yearIndex(x)
}
\arguments{
\item{x}{A vector or column of dates}
}
\description{
Function to create a year index variable
}
\examples{
# given a vector of dates
df <- data.frame(date = as.Date(c("2017-01-01","2018-04-01","2019-07-01","2019-08-01"),
format = "\%Y-\%m-\%d"))
# calculate the time period variable
df$period <- yearIndex(df$date)
df
}
