% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gk.R
\name{GKIndex}
\alias{GKIndex}
\title{Compute the Geary-Khamis index}
\usage{
GKIndex(x, pvar, qvar, pervar, prodID, sample = "", window, splice = "mean")
}
\arguments{
\item{x}{A dataframe containing price, quantity, a time period identifier
and a product identifier. It must have column names.}

\item{pvar}{A character string for the name of the price variable}

\item{qvar}{A character string for the name of the quantity variable}

\item{pervar}{A character string for the name of the time variable. This variable
must contain integers starting at period 1 and increasing in increments of 1 period.
There may be observations on multiple products for each time period.}

\item{prodID}{A character string for the name of the product identifier}

\item{sample}{set to "matched" to only use products that occur
across all periods in a given window. Default is not to match.}

\item{window}{An integer specifying the length of the window.}

\item{splice}{the splicing method to use to extend the index. Valid methods are
window, movement, half or mean. The default is mean.}
}
\description{
Compute the Geary-Khamis index
}
\examples{
# compute a Geary-Khamis index with mean splicing
GKIndex(CES_sigma_2, pvar = "prices", qvar = "quantities", pervar = "time",
prodID = "prodID", window=11, splice = "mean")
}
\references{
Ivancic, L., W.E. Diewert and K.J. Fox (2011), "Scanner Data,
Time Aggregation and the Construction of Price Indexes", Journal of
Econometrics 161, 24-35.
}
