% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dissimilarity.R
\name{maximiumSimilarityLinks}
\alias{maximiumSimilarityLinks}
\title{function to compute the maximum similarity chain links from
a measure of dissimilarity}
\usage{
maximiumSimilarityLinks(x)
}
\arguments{
\item{x}{a matrix containing a dissimilarity measure where
the first two columns are the indices and the third column
is the dissimilarity measure.}
}
\description{
function to compute the maximum similarity chain links from
a measure of dissimilarity
}

