% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dissimilarity.R
\name{mixScaleDissimilarity}
\alias{mixScaleDissimilarity}
\title{mixScaleDissimilarity}
\usage{
mixScaleDissimilarity(x, pvar, qvar, prodID, pervar, measure = "absolute",
  combine = "geomean")
}
\arguments{
\item{x}{A dataframe}

\item{pvar}{string identifying the price variable in x}

\item{qvar}{string identifying the quantity variable in x}

\item{prodID}{string identifying the product id variable in x}

\item{pervar}{string identifying the time period variable in x}

\item{measure}{choice of dissimilarity measure. Valid options
are mix, scale or absolute.}

\item{combine}{specifies how to combine the price and quantity vectors.
"stack" stacks the price and quantity vectors, "geomean" computes
separate dissimilarity measures for prices and quantities then takes
the geometric mean of these.}
}
\value{
A matrix where the first two columns are the possible combinations
of periods and the third column is the dissimilarity measure.
}
\description{
Fox, Hill and Diewert 2004 vector dissimilarity
}
