% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitValues.R
\name{monthIndex}
\alias{monthIndex}
\title{Generate an index of months}
\usage{
monthIndex(x)
}
\arguments{
\item{x}{A vector or column of dates}
}
\description{
A function to create a month index variable
}
