% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitValues.R
\name{quarterIndex}
\alias{quarterIndex}
\title{Generate an index of quarters}
\usage{
quarterIndex(x)
}
\arguments{
\item{x}{A vector or column of dates}
}
\description{
A function to create a quarter index variable
}
\examples{
# given a vector of dates
df <- data.frame(date = as.Date(c("2017-01-01","2017-04-01","2017-07-01","2017-08-01"),
format = "\%Y-\%m-\%d"))
# calculate the time period variable
df$period <- quarterIndex(df$date)
df
}
