% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{quarterIndex}
\alias{quarterIndex}
\title{quarterIndex}
\usage{
quarterIndex(x)
}
\arguments{
\item{x}{A vector or column of dates}
}
\description{
A function to create a quarter index variable
}
\keyword{internal}

