% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator.R
\name{volumeIndicator}
\alias{volumeIndicator}
\title{Compute a volume indicator}
\usage{
volumeIndicator(x, pvar, qvar, pervar, prodID, method,
  sample = "matched")
}
\arguments{
\item{x}{data frame with input data}

\item{pvar}{character string for the name of the price column}

\item{qvar}{character string for the name of the quantity column}

\item{pervar}{character string for the name of the time period variable}

\item{prodID}{character string for the name of the product ID column}

\item{method}{character string for the indicator method. Valid options
are "laspeyres", "paasche", "bennet", or "montgomery".}

\item{sample}{whether to use a matched sample (sample = "matched")}
}
\value{
an nx1 matrix containing the indicator
}
\description{
This computes an indicator of the change in volume,
using the differences approach to measuring change.
}
