% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtpd.R
\name{WTPDIndex}
\alias{WTPDIndex}
\title{Compute a weighted time-product-dummy multilateral index}
\usage{
WTPDIndex(
  x,
  pvar,
  qvar,
  pervar,
  prodID,
  sample = "",
  window = 13,
  splice = "mean"
)
}
\arguments{
\item{x}{A dataframe containing price, quantity, a time period identifier
and a product identifier. It must have column names.}

\item{pvar}{A character string for the name of the price variable}

\item{qvar}{A character string for the name of the quantity variable}

\item{pervar}{A character string for the name of the time variable. This variable
must contain integers starting at period 1 and increasing in increments of 1 period.
There may be observations on multiple products for each time period.}

\item{prodID}{A character string for the name of the product identifier}

\item{sample}{set to "matched" to only use products that occur
across all periods in a given window. Default is not to match.}

\item{window}{An integer specifying the length of the window.}

\item{splice}{A character string specifying the splicing method. Valid methods are
window, movement, half or mean. The default is mean.}
}
\description{
A function to calculate a weighted-time-product-dummy multilateral index.
}
\details{
When there are missing values in the dataset (e.g., from new or disappearing
products), the default option is to treat the missing prices and quantities
as zero. An alternative is to use a matched sample, where only products that
appear throughout each window in the calculation are kept.
}
\examples{
# compute a wtpd index with mean splicing
WTPDIndex(CES_sigma_2, pvar = "prices", qvar = "quantities", pervar = "time",
prodID = "prodID", window=11, splice = "mean")
}
\references{
Ivancic, L., W.E. Diewert and K.J. Fox (2011), "Scanner Data,
Time Aggregation and the Construction of Price Indexes", Journal of
Econometrics 161, 24-35.
}
