% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{yearOverYearIndexes}
\alias{yearOverYearIndexes}
\title{Estimate year-over-year indexes}
\usage{
yearOverYearIndexes(freq, indexFunction, indexArgs)
}
\arguments{
\item{freq}{the frequency of the data. Either "monthly" or "quarterly".}

\item{indexFunction}{the function to use to calculate the index. Available
options are `priceIndex`, `GEKSIndex`, `GKIndex`, `WTPDIndex`.}

\item{indexArgs}{arguments for the price index function as a named list.
All arguments must be named.}
}
\value{
a list of indexes with one element for each month or quarter
}
\description{
Estimate year-over-year indexes
}
